Переписал шапку сайта с использованием NavLink
