+++ src/components/AppHeader/AppHeader.module.css
@@ -28,3 +28,26 @@
     display: flex;
     justify-content: flex-end;
 }
+
+.navLink {
+    height: 50px;
+    display: flex;
+    align-items: center;
+    border-radius: 40px;
+    color: #8585ad;
+    cursor: pointer;
+
+    background: none;
+    border-style: hidden;
+    text-decoration: none;
+}
+
+.isActive {
+    color: #f2f2f3;
+}
+
+@media screen and (max-width: 1200px) {
+    .text {
+        display: none;
+    }
+}
