--- src/components/AppHeader/NavLink/NavLink.module.css
@@ -1,21 +1,0 @@
-.wrapper {
-    height: 50px;
-    display: flex;
-    align-items: center;
-    border-radius: 40px;
-    color: #8585ad;
-    cursor: pointer;
-
-    background: none;
-    border-style: hidden;
-}
-
-.isActive {
-    color: #f2f2f3;
-}
-
-@media screen and (max-width: 1200px) {
-    .text {
-        display: none;
-    }
-}
