--- src/components/AppHeader/NavLink/index.ts
@@ -1,1 +1,0 @@
-export * from './NavLink';
